WCM-573 Extract methods to base class since they'll be used in another display context
