+++ content-targeting-web/src/main/java/com/liferay/content/targeting/web/display/context/ContentTargetingEditCampaignDisplayContext.java
@@ -15,62 +15,43 @@
 package com.liferay.content.targeting.web.display.context;
 
 import com.liferay.content.targeting.model.Campaign;
-import com.liferay.content.targeting.model.UserSegment;
-import com.liferay.content.targeting.service.CampaignLocalServiceUtil;
 import com.liferay.content.targeting.service.UserSegmentLocalService;
 import com.liferay.content.targeting.util.ContentTargetingUtil;
 import com.liferay.content.targeting.util.UserSegmentUtil;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
 import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.JavaConstants;
 import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.PortalUtil;
-import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.TimeZoneUtil;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.kernel.util.WebKeys;
 
-import java.util.Calendar;
-import java.util.List;
-
-import javax.portlet.PortletConfig;
 import javax.portlet.PortletURL;
-
-import javax.servlet.http.HttpServletRequest;
 
 /**
  * @author Jürgen Kappler
  */
-public class ContentTargetingEditCampaignDisplayContext {
+public class ContentTargetingEditCampaignDisplayContext
+	extends BaseContentTargetingCampaignDisplayContext {
 
 	public ContentTargetingEditCampaignDisplayContext(
 		LiferayPortletRequest liferayPortletRequest,
 		LiferayPortletResponse liferayPortletResponse,
 		UserSegmentLocalService userSegmentLocalService) {
 
-		_liferayPortletRequest = liferayPortletRequest;
-		_liferayPortletResponse = liferayPortletResponse;
-		_userSegmentLocalService = userSegmentLocalService;
-
-		_request = PortalUtil.getHttpServletRequest(_liferayPortletRequest);
-
-		_themeDisplay = (ThemeDisplay)_request.getAttribute(
-			WebKeys.THEME_DISPLAY);
+		super(
+			liferayPortletRequest, liferayPortletResponse,
+			userSegmentLocalService);
 	}
 
 	public String getBackURL() {
-		String backURL = ParamUtil.getString(_request, "backURL");
+		String backURL = ParamUtil.getString(request, "backURL");
 
 		if (Validator.isNotNull(backURL)) {
 			return backURL;
 		}
 
-		PortletURL backURLObject = _liferayPortletResponse.createRenderURL();
+		PortletURL backURLObject = liferayPortletResponse.createRenderURL();
 
 		backURLObject.setParameter("mvcPath", "/view.jsp");
 		backURLObject.setParameter("tabs1", "campaigns");
@@ -78,100 +59,16 @@ public class ContentTargetingEditCampaignDisplayContext {
 		return backURLObject.toString();
 	}
 
-	public Campaign getCampaign() {
-		if (_campaign != null) {
-			return _campaign;
-		}
-
-		long campaignId = getCampaignId();
-
-		if (campaignId > 0) {
-			_campaign = CampaignLocalServiceUtil.fetchCampaign(campaignId);
-		}
-
-		return _campaign;
-	}
-
-	public long getCampaignId() {
-		if (_campaignId != null) {
-			return _campaignId;
-		}
-
-		_campaignId = ParamUtil.getLong(_request, "campaignId", 0);
-
-		return _campaignId;
-	}
-
-	public String getCampaignTitle() {
-		if (Validator.isNotNull(_campaignTitle)) {
-			return _campaignTitle;
-		}
-
-		Campaign campaign = getCampaign();
-
-		if (campaign != null) {
-			_campaignTitle = campaign.getName(_themeDisplay.getLocale());
-		}
-		else {
-			PortletConfig portletConfig =
-				(PortletConfig) _liferayPortletRequest.getAttribute(
-					JavaConstants.JAVAX_PORTLET_CONFIG);
-
-			_campaignTitle = LanguageUtil.get(
-				portletConfig.getResourceBundle(_themeDisplay.getLocale()),
-				"new-campaign");
-		}
-
-		return _campaignTitle;
-	}
-
-	public Calendar getEndDate() {
-		if (_endDate != null) {
-			return _endDate;
-		}
-
-		Campaign campaign = getCampaign();
-
-		Calendar endDate = Calendar.getInstance();
-
-		if (campaign != null) {
-			endDate.setTime(campaign.getEndDate());
-		}
-		else {
-			endDate.add(Calendar.YEAR, 1);
-		}
-
-		_endDate = endDate;
-
-		return _endDate;
-	}
-
-	public Integer getPriority() {
-		if (_priority != null) {
-			return _priority;
-		}
-
-		_priority = 1;
-
-		Campaign campaign = getCampaign();
-
-		if (campaign != null) {
-			_priority = campaign.getPriority();
-		}
-
-		return _priority;
-	}
-
 	public String getRedirect() {
 		if (Validator.isNotNull(_redirect)) {
 			return _redirect;
 		}
 
-		String redirect = ParamUtil.getString(_request, "redirect");
+		String redirect = ParamUtil.getString(request, "redirect");
 
 		if (Validator.isNull(redirect)) {
 			PortletURL redirectURLObject =
-				_liferayPortletResponse.createRenderURL();
+				liferayPortletResponse.createRenderURL();
 
 			redirectURLObject.setParameter("mvcPath", "/view.jsp");
 			redirectURLObject.setParameter("tabs1", "campaigns");
@@ -184,96 +81,6 @@ public class ContentTargetingEditCampaignDisplayContext {
 		return _redirect;
 	}
 
-	public Calendar getStartDate() {
-		if (_startDate != null) {
-			return _startDate;
-		}
-
-		Campaign campaign = getCampaign();
-
-		Calendar startDate = Calendar.getInstance();
-
-		if (campaign != null) {
-			startDate.setTime(campaign.getEndDate());
-		}
-
-		_startDate = startDate;
-
-		return _startDate;
-	}
-
-	public String getTimeZoneId() {
-		if (Validator.isNotNull(_timeZoneId)) {
-			return _timeZoneId;
-		}
-
-		_timeZoneId = TimeZoneUtil.getDefault().getID();
-
-		Campaign campaign = getCampaign();
-
-		if (campaign != null) {
-			_timeZoneId = campaign.getTimeZoneId();
-		}
-
-		return _timeZoneId;
-	}
-
-	public String getUserSegmentAssetCategoryIdsAsString() {
-		if (Validator.isNotNull(_userSegmentAssetCategoryIdsAsString)) {
-			return _userSegmentAssetCategoryIdsAsString;
-		}
-
-		String userSegmentAssetCategoryIdsAsString = StringPool.BLANK;
-
-		long campaignId = getCampaignId();
-
-		if (campaignId > 0) {
-			List<UserSegment> campaignUserSegments = null;
-
-			campaignUserSegments =
-				_userSegmentLocalService.getCampaignUserSegments(campaignId);
-
-			long[] userSegmentAssetCategoryIds =
-				ContentTargetingUtil.getAssetCategoryIds(campaignUserSegments);
-
-			userSegmentAssetCategoryIdsAsString = StringUtil.merge(
-				userSegmentAssetCategoryIds);
-		}
-
-		_userSegmentAssetCategoryIdsAsString =
-			userSegmentAssetCategoryIdsAsString;
-
-		return _userSegmentAssetCategoryIdsAsString;
-	}
-
-	public String getUserSegmentAssetCategoryNames() {
-		if (Validator.isNotNull(_userSegmentAssetCategoryNames)) {
-			return _userSegmentAssetCategoryNames;
-		}
-
-		String userSegmentAssetCategoryNames = StringPool.BLANK;
-
-		long campaignId = getCampaignId();
-
-		if (campaignId > 0) {
-			List<UserSegment> campaignUserSegments = null;
-
-			campaignUserSegments =
-				_userSegmentLocalService.getCampaignUserSegments(campaignId);
-
-			long[] userSegmentAssetCategoryIds =
-				ContentTargetingUtil.getAssetCategoryIds(campaignUserSegments);
-
-			userSegmentAssetCategoryNames =
-				ContentTargetingUtil.getAssetCategoryNames(
-					userSegmentAssetCategoryIds, _themeDisplay.getLocale());
-		}
-
-		_userSegmentAssetCategoryNames = userSegmentAssetCategoryNames;
-
-		return _userSegmentAssetCategoryNames;
-	}
-
 	public String getVocabularyGroupIds() throws PortalException {
 		if (Validator.isNotNull(_vocabularyGroupIds)) {
 			return _vocabularyGroupIds;
@@ -281,15 +88,15 @@ public class ContentTargetingEditCampaignDisplayContext {
 
 		long[] vocabularyGroupIds = new long[1];
 
-		if (_themeDisplay.getScopeGroupId() ==
-				_themeDisplay.getCompanyGroupId()) {
+		if (themeDisplay.getScopeGroupId() ==
+				themeDisplay.getCompanyGroupId()) {
 
-			vocabularyGroupIds[0] = _themeDisplay.getCompanyGroupId();
+			vocabularyGroupIds[0] = themeDisplay.getCompanyGroupId();
 		}
 		else {
 			vocabularyGroupIds =
 				ContentTargetingUtil.getAncestorsAndCurrentGroupIds(
-					_themeDisplay.getSiteGroupId());
+					themeDisplay.getSiteGroupId());
 		}
 
 		_vocabularyGroupIds = StringUtil.merge(vocabularyGroupIds);
@@ -304,15 +111,15 @@ public class ContentTargetingEditCampaignDisplayContext {
 
 		long[] vocabularyIds = new long[1];
 
-		if (_themeDisplay.getScopeGroupId() ==
-				_themeDisplay.getCompanyGroupId()) {
+		if (themeDisplay.getScopeGroupId() ==
+				themeDisplay.getCompanyGroupId()) {
 
 			ServiceContext serviceContext = new ServiceContext();
 
-			serviceContext.setScopeGroupId(_themeDisplay.getScopeGroupId());
+			serviceContext.setScopeGroupId(themeDisplay.getScopeGroupId());
 
 			vocabularyIds[0] = UserSegmentUtil.getAssetVocabularyId(
-				_themeDisplay.getUserId(), serviceContext);
+				themeDisplay.getUserId(), serviceContext);
 		}
 		else {
 			long[] vocabularyGroupsIds = StringUtil.split(
@@ -337,21 +144,7 @@ public class ContentTargetingEditCampaignDisplayContext {
 		return true;
 	}
 
-	private Campaign _campaign;
-	private Long _campaignId;
-	private String _campaignTitle;
-	private Calendar _endDate;
-	private final LiferayPortletRequest _liferayPortletRequest;
-	private final LiferayPortletResponse _liferayPortletResponse;
-	private Integer _priority;
 	private String _redirect;
-	private final HttpServletRequest _request;
-	private Calendar _startDate;
-	private final ThemeDisplay _themeDisplay;
-	private String _timeZoneId;
-	private String _userSegmentAssetCategoryIdsAsString;
-	private String _userSegmentAssetCategoryNames;
-	private final UserSegmentLocalService _userSegmentLocalService;
 	private String _vocabularyGroupIds;
 	private String _vocabularyIds;
